SLight rearangement of TextPollBlockViewModel and TextPollBlockView
